requirements.txt
asgiref==3.7.2
Django==4.2.2
django-debug-toolbar==4.2.0
django-js-asset==2.2.0
django-mptt==0.15.0
django-mptt-admin==2.6.1
django-taggit==3.0.0
Pillow==10.1.0
pytils==0.4.1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==4.2.2
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.2.0
django-mptt==0.15.0
django-mptt-admin==2.6.1
django-recaptcha==4.0.0
django-taggit==3.0.0
Pillow==10.1.0
pytils==0.4.1
sqlparse==0.4.4
typing_extensions==4.9.0
